Removing all usize ids
- Added absolute read limit field to `ConfigLoader` in the case of an increase of `READ_LIMIE`.
- Changed `ModuleId` to u32 and `ScopeId` to u16. Reason being, for something like `TypeInfo`, there is a net 8 bytes added to to every single type, which is a negative for cache lining. The compiler itself generates 40 type symbols which already makes a 320 -> 160 byte difference. `ScopeId` is u16 since on average only one `.chrn` file has more than 3 scopes in it, to where I doubt 30K (or any at all) chrn files will be in one given compilation session.

+++ chrn_core/compilation/src/modules.rs
@@ -644,6 +644,6 @@ fn resolve_modules(
 
         // Needs - 1 so that it fits inside the temporary Vec, but still uses it's actual module
         // id.
-        other_mods[current_mod_id.id - 1] = Some(sub_mod);
+        other_mods[(current_mod_id.id - 1) as usize] = Some(sub_mod);
     }
 }
